+++ bot/models/user.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from enum import Enum as PyEnum
 
-from sqlalchemy import Boolean, Column, Date, DateTime, Enum, Integer, orm, String
+from sqlalchemy import BigInteger, Boolean, Column, Date, DateTime, Enum, Integer, orm, String
 
 from database import Base
 
@@ -14,7 +14,7 @@ class LanguageEnum(PyEnum):
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     username = Column(String, unique=True, nullable=False)
     token_count = Column(Integer, default=0)
     token_last_receipt = Column(Date, nullable=True)
